we need to keep track of the last tick of the change in modes of the ghosts

+++ src/main.cpp
@@ -6,8 +6,8 @@
 int main() {
 	using namespace std::this_thread; // sleep_for, sleep_until
 	using namespace std::chrono; // nanoseconds, system_clock, seconds
-	float ticksPerSecond = 5;
-	milliseconds timeout((int)((1 / ticksPerSecond) * 1000));
+	int ticksPerSecond = 5;
+	milliseconds timeout((int)((1.0 / ticksPerSecond) * 1000));
 	Game game = Game("../maps/0.txt", ticksPerSecond, 1, 3);
 	/*
 	Current solution only works for windows! (the GetAsyncKeyState function)
